Fixed postcondition violation in ES_TOOL_FOUNDATIONS.foundation_widget. Fixed bug which initialized services on recycle when the tool was not initialized.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@73473 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/Eiffel/interface/graphical/foundations/dialogs/es_dialog.e
@@ -182,7 +182,7 @@ feature {NONE} -- Clean up
 
 				button_actions.wipe_out
 
-				if internal_dialog /= Void then
+				if internal_dialog /= Void and then not internal_dialog.is_destroyed then
 					internal_dialog.hide
 					internal_dialog.destroy
 				end
